\name{@ANNOBJPREFIX@PROSITE}
\alias{@ANNOBJPREFIX@PROSITE}
\title{Maps between Manufacturer Identifiers and PROSITE Identifiers}
\description{
   @ANNOBJPREFIX@PROSITE is an R object that provides mappings between
   manufacturer identifiers and PROSITE identifiers.
}

\details{
The bimap interface for PROSITE is defunct. Please use select() interface to
PROSITE identifiers. See ?AnnotationDbi::select for details.
}

\keyword{datasets}
